+++ src/test/scala/io/iohk/ethereum/consensus/pow/validators/RestrictedEthashBlockHeaderValidatorSpec.scala
@@ -94,7 +94,7 @@ class RestrictedEthashBlockHeaderValidatorSpec
         daoForkConfig = None,
         // unused
         maxCodeSize = None,
-        chainId = 0x3d.toByte,
+        chainId = 0x3d,
         networkId = 1,
         monetaryPolicyConfig = null,
         customGenesisFileOpt = None,
