+++ src/test/scala/io/iohk/ethereum/jsonrpc/JsonRpcControllerEthTransactionSpec.scala
@@ -575,7 +575,7 @@ class JsonRpcControllerEthTransactionSpec
           value = 0,
           payload = ByteString()
         ),
-        signature = ECDSASignature(0, 0, 0.toByte),
+        signature = ECDSASignature(0, 0, 0),
         sender = Address("0x1234")
       )
       PendingTransaction(fakeTransaction, System.currentTimeMillis)
