+++ src/test/scala/io/iohk/ethereum/jsonrpc/MantisServiceSpec.scala
@@ -52,7 +52,7 @@ class MantisServiceSpec
             value = 0,
             payload = ByteString()
           ),
-          signature = ECDSASignature(0, 0, 0.toByte),
+          signature = ECDSASignature(0, 0, 0),
           sender = Address("0x1234")
         )
