+++ src/test/scala/io/iohk/ethereum/network/p2p/messages/TransactionSpec.scala
@@ -31,7 +31,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
 
   val validTransactionSignatureOldSchema = SignedTransaction(
     validTx,
-    pointSign = 28.toByte,
+    pointSign = 28,
     signatureRandom = ByteString(Hex.decode("cfe3ad31d6612f8d787c45f115cc5b43fb22bcc210b62ae71dc7cbf0a6bea8df")),
     signature = ByteString(Hex.decode("57db8998114fae3c337e99dbd8573d4085691880f4576c6c1f6c5bbfe67d6cf0")),
     chainId = blockchainConfig.chainId
@@ -39,7 +39,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
 
   val invalidTransactionSignatureNewSchema = SignedTransaction(
     validTx,
-    pointSign = (-98).toByte,
+    pointSign = -98,
     signatureRandom = ByteString(Hex.decode("cfe3ad31d6612f8d787c45f115cc5b43fb22bcc210b62ae71dc7cbf0a6bea8df")),
     signature = ByteString(Hex.decode("57db8998114fae3c337e99dbd8573d4085691880f4576c6c1f6c5bbfe67d6cf0")),
     chainId = blockchainConfig.chainId
@@ -47,7 +47,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
 
   val invalidStx = SignedTransaction(
     validTx.copy(gasPrice = 0),
-    pointSign = (-98).toByte,
+    pointSign = -98,
     signatureRandom = ByteString(Hex.decode("cfe3ad31d6612f8d787c45f115cc5b43fb22bcc210b62ae71dc7cbf0a6bea8df")),
     signature = ByteString(Hex.decode("57db8998114fae3c337e99dbd8573d4085691880f4576c6c1f6c5bbfe67d6cf0")),
     chainId = blockchainConfig.chainId
@@ -71,7 +71,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
 
   val validSignedTransactionForNewSigningScheme = SignedTransaction(
     tx = validTransactionForNewSigningScheme,
-    pointSign = (-98).toByte,
+    pointSign = -98,
     signatureRandom = ByteString(Hex.decode("1af423b3608f3b4b35e191c26f07175331de22ed8f60d1735f03210388246ade")),
     signature = ByteString(Hex.decode("4d5b6b9e3955a0db8feec9c518d8e1aae0e1d91a143fbbca36671c3b89b89bc3")),
     blockchainConfig.chainId
@@ -79,7 +79,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
 
   val stxWithInvalidPointSign = SignedTransaction(
     validTx,
-    pointSign = 26.toByte,
+    pointSign = 26,
     signatureRandom = ByteString(Hex.decode("cfe3ad31d6612f8d787c45f115cc5b43fb22bcc210b62ae71dc7cbf0a6bea8df")),
     signature = ByteString(Hex.decode("57db8998114fae3c337e99dbd8573d4085691880f4576c6c1f6c5bbfe67d6cf0")),
     blockchainConfig.chainId
@@ -115,7 +115,7 @@ class TransactionSpec extends AnyFlatSpec with Matchers {
         value = BigInt(31337),
         payload = ByteString.empty
       ),
-      pointSign = 28.toByte,
+      pointSign = 28,
       signatureRandom =
         ByteString(BigInt("61965845294689009770156372156374760022787886965323743865986648153755601564112").toByteArray),
       signature =
